Former-commit-id: ab617839e9da6d7576bb2ee260aa5a53691042a0 [formerly 0b7b354ef529ca3b9795f8042772890125942701] [formerly 1390383adba6cfc347d513cc290927a6d254d733 [formerly 6a6a1a4f56d26a6601d8a5e53a6fc99dbcc187be]]
Former-commit-id: 1b19d5fda60c06437b1c874c424258a83dad0f77 [formerly 2423a7b5c83f99d5097c3cf5dc647033161ad78a]
Former-commit-id: 744ec2d55571319ba6266ac69b6a51ba8deb0b4a

+++ include/Nazara/Audio/Music.hpp
@@ -48,7 +48,7 @@ namespace Nz
 			UInt32 GetDuration() const;
 			AudioFormat GetFormat() const;
 			UInt32 GetPlayingOffset() const;
-			UInt32 GetSampleCount() const;
+			UInt64 GetSampleCount() const;
 			UInt32 GetSampleRate() const;
 			SoundStatus GetStatus() const;
